--- design-system/theme/README.md
@@ -1,9 +1,0 @@
----
-title: Theme
-path: '/design-system/theme'
-type: design-system
----
-
-# Theme
-
-This is our theme
