+++ design-system/theme/README.mdx
@@ -1,0 +1,69 @@
+---
+title: Theme
+path: '/design-system/theme'
+type: design-system
+---
+
+import { ColorsExample } from './colorsExample';
+
+# Theme
+
+This is our theme
+
+## Colors
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+We use semantic colours
+
+<ColorsExample />
+
+## fontFamilies
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+## spacing
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+## breakpoints
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+## borderRadius
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+## boxShadow
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+## fontsizes
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+## baseHeadingStyle
+
+```jsx
+import { colors } from 'design-system/theme';
+```
+
+## headings
+
+```jsx
+import { colors } from 'design-system/theme';
+```
